requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
openai==1.54.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
pylogram==0.12.3

duckduckgo_search
beautifulsoup4

# Dashboard
nicegui>=2.0.0
plotly>=5.18.0
pyvis>=0.3.2
bcrypt>=4.1.0
